+++ templates/export.tf
@@ -5,14 +5,21 @@ variable "export_vars_to_scalr" {
 
 variable "scalr_workspace_id" {
   type    = string
+  default = null
+}
+
+variable "scalr_var_set_id" {
+  type    = string
+  default = null
 }
 
 data "external" "example" {
   program = ["python3", "${path.module}/migrate-environment.py"]
 
   query = {
-    variables = join(",", var.export_vars_to_scalr)
-    workspace_id = var.scalr_workspace_id
+    variables    = join(",", var.export_vars_to_scalr)
+    workspace_id = var.scalr_workspace_id != null ? var.scalr_workspace_id : ""
+    var_set_id   = var.scalr_var_set_id != null ? var.scalr_var_set_id : ""
   }
 }